Hackathon_Project/requirements.txt
flask
python-dotenv
google-generativeai>=0.5.0
requests
descope==1.7.9
